Moving things around, refactor begins
